# foundry_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AcadAssist API",
    "description": "Backend API powering the AcadAssist AI Study Assistant via Microsoft Foundry.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/tools/search_knowledge": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Search student academic knowledge base",
        "description": "Search academic materials using hybrid retrieval. Requires user_id and query.",
        "operationId": "search_knowledge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchKnowledgeParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchKnowledgeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/summarize_document": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Summarize an academic document",
        "description": "Summarize a specific academic document or textbook chapter in the knowledge base.",
        "operationId": "summarize_document",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizeDocumentParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Summarize Document"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/generate_quiz": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Generate an practice quiz",
        "description": "Generate practice quiz questions on a specified academic subject or topic.",
        "operationId": "generate_quiz",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuizParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Generate Quiz"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/submit_quiz": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Submit student quiz answers",
        "description": "Submit answers for a previously generated quiz, calculate score, and record history.",
        "operationId": "submit_quiz",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Submit Quiz"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/get_performance": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Retrieve historical student performance metrics",
        "description": "Retrieve historical assessment scores, quiz statistics, and performance trends.",
        "operationId": "get_performance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPerformanceParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Performance"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/get_weak_topics": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Identify student weak topics",
        "description": "Identify topics where the student has struggled based on quiz and assessment history.",
        "operationId": "get_weak_topics",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetWeakTopicsParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Weak Topics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/get_progress": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Retrieve course completion progress",
        "description": "Retrieve current course completion percentage, reading milestone progress, and study streaks.",
        "operationId": "get_progress",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetProgressParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Progress"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/get_upcoming_exams": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Retrieve upcoming exams and deadlines",
        "description": "Retrieve scheduled exams, tests, and assignment deadlines.",
        "operationId": "get_upcoming_exams",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUpcomingExamsParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Upcoming Exams"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/create_study_plan": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Create an optimized study plan",
        "description": "Create an optimized daily study plan tailored to weak topics and available study hours.",
        "operationId": "create_study_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudyPlanParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Study Plan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/get_today_plan": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Retrieve today's study schedule",
        "description": "Retrieve the student's scheduled study sessions and practice tasks for today.",
        "operationId": "get_today_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTodayPlanParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Today Plan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/generate_weekly_report": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Generate a weekly learning summary",
        "description": "Generate a weekly learning summary including hours studied, concepts mastered, and streak status.",
        "operationId": "generate_weekly_report",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWeeklyReportParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Generate Weekly Report"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateStudyPlanParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "description": "Start date (YYYY-MM-DD)"
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "description": "End date (YYYY-MM-DD)"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "start_date",
          "end_date"
        ],
        "title": "CreateStudyPlanParams"
      },
      "GenerateQuizParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "topic_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {}
            ],
            "title": "Topic Ids"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "description": "Quiz difficulty (easy, medium, hard)"
          },
          "count": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Count",
            "description": "Number of questions",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "subject_id",
          "difficulty"
        ],
        "title": "GenerateQuizParams"
      },
      "GenerateWeeklyReportParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "week_start": {
            "type": "string",
            "title": "Week Start",
            "description": "Week start date (YYYY-MM-DD)"
          },
          "week_end": {
            "type": "string",
            "title": "Week End",
            "description": "Week end date (YYYY-MM-DD)"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "week_start",
          "week_end"
        ],
        "title": "GenerateWeeklyReportParams"
      },
      "GetPerformanceParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "GetPerformanceParams"
      },
      "GetProgressParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "course_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "GetProgressParams"
      },
      "GetTodayPlanParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "GetTodayPlanParams"
      },
      "GetUpcomingExamsParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "GetUpcomingExamsParams"
      },
      "GetWeakTopicsParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "GetWeakTopicsParams"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SearchKnowledgeParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "course_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "query"
        ],
        "title": "SearchKnowledgeParams"
      },
...
      "SearchKnowledgeResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ChunkResult"
            },
            "type": "array",
            "title": "Results",
            "description": "List of matching chunks"
          }
        },
        "type": "object",
        "title": "SearchKnowledgeResponse",
        "description": "Response payload containing retrieved knowledge chunks."
      },
...
      "SubmitQuizParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "quiz_id": {
            "type": "string",
            "title": "Quiz Id"
          },
          "answers": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {}
            ],
            "title": "Answers",
            "description": "Student's chosen answers"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "quiz_id",
          "answers"
        ],
        "title": "SubmitQuizParams"
      },
      "SummarizeDocumentParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Summary mode (e.g. concise, detailed, bullet_points)"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "document_id",
          "mode"
        ],
        "title": "SummarizeDocumentParams"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://lqs5ss0s-8000.inc1.devtunnels.ms"
    }
  ]
}